split spatial_compare and add notebook directory
