--- __init__.py
@@ -1,1 +1,0 @@
-from .spatial_compare import SpatialCompare, spatial_detection_scores, compare_reference_and_spatial,summarize_and_plot
