+++ spatial_compare/__init__.py
@@ -1,3 +1,2 @@
 from .spatial_compare import SpatialCompare, get_column_ordering
-
-
+from .utils import grouped_obs_mean,spatial_detection_scores,summarize_and_plot, compare_reference_and_spatial
